dwm: Add in the other tags
I want this to be somewhat similar to the number of tags I use in i3wm
on my other computer, so we're going with 10 tags, meaning that the '0'
key has to be moved away from showing everything, and since MODKEY+a is
not bound, and 'a' kind of means 'all', we'll use that for showing and
tagging all. Also use ControlMask instead of ShiftMask for the tag all
command.